Move `null` package under `pantherlog`
