+++ internal/log_analysis/log_processor/pantherlog/enums.go
@@ -36,7 +36,7 @@ import (
 
 	jsoniter "github.com/json-iterator/go"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/common/null"
+	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/null"
 )
 
 // Enum maps numbers to labels
