+++ internal/log_analysis/log_processor/pantherlog/enums_test.go
@@ -34,7 +34,7 @@ import (
 	jsoniter "github.com/json-iterator/go"
 	"github.com/stretchr/testify/require"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/common/null"
+	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/null"
 )
 
 func TestNewEnumDecoder(t *testing.T) {
