+++ internal/log_analysis/log_processor/pantherlog/null/int64.go
