+++ internal/log_analysis/log_processor/pantherlog/null/null.go
