+++ internal/log_analysis/log_processor/pantherlog/null/string.go
