+++ internal/log_analysis/log_processor/pantherlog/null/uint16.go
