+++ internal/log_analysis/log_processor/pantherlog/null/uint32.go
