+++ internal/log_analysis/log_processor/pantherlog/validator.go
@@ -23,7 +23,7 @@ import (
 
 	"gopkg.in/go-playground/validator.v9"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/common/null"
+	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/null"
 )
 
 var (
