+++ internal/log_analysis/log_processor/parsers/parsers.go
@@ -26,8 +26,8 @@ import (
 	jsoniter "github.com/json-iterator/go"
 	"gopkg.in/go-playground/validator.v9"
 
-	"github.com/panther-labs/panther/internal/log_analysis/log_processor/common/null"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog"
+	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/null"
 	"github.com/panther-labs/panther/internal/log_analysis/log_processor/pantherlog/rowid"
 )
